move stuff from page to component; fix style regression
